+++ lib/src/http.dart
@@ -1,7 +1,6 @@
 import 'dart:convert';
 
 import 'package:http/http.dart' as http;
-import 'package:logger/logger.dart';
 
 class RequestTemplate {
   Map<String, String> headers = <String, String>{};
@@ -51,7 +50,6 @@ abstract class HttpRequestResponse {
 
 class Http {
   static final Http _singleton = Http._internal();
-  final Logger _logger = Logger();
 
   List<HttpInterceptor> interceptors = [HttpJsonInterceptor()];